+++ src/ConsoleColour/Core/Contracts/TerminalInterface.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace AlecRabbit\ConsoleColour\Contracts;
+namespace AlecRabbit\ConsoleColour\Core\Contracts;
 
 /**
  * Interface TerminalInterface
@@ -22,12 +22,19 @@ interface TerminalInterface
     public function height(bool $recheck = false): int;
 
     /**
+     * @param bool $recheck
      * @return bool
      */
-    public function supports256Color(): bool;
+    public function supports256Color(bool $recheck = false): bool;
 
     /**
+     * @param bool $recheck
      * @return bool
      */
-    public function supportsColor(): bool;
+    public function supportsColor(bool $recheck = false): bool;
+
+    /**
+     * @param string $title
+     */
+    public function setTitle(string $title): void;
 }