+++ tests/Unit/ThemesTest.php
@@ -2,6 +2,7 @@
 
 namespace AlecRabbit\Tests\ConsoleColour;
 
+use AlecRabbit\ConsoleColour\ConsoleColor;
 use AlecRabbit\ConsoleColour\Exception\InvalidStyleException;
 use AlecRabbit\ConsoleColour\Themes;
 use AlecRabbit\Tests\Helper;
@@ -9,27 +10,45 @@ use PHPUnit\Framework\TestCase;
 
 class ThemesTest extends TestCase
 {
+    public const ESC = ConsoleColor::ESC_CHAR;
+
     public const THEMES = [
-        'italic' => "\033[3m",
-        'dark' => "\033[2m",
-        'bold' => "\033[1m",
-        'darkItalic' => "\033[2;3m",
-        'white' => "\033[97m",
-        'whiteBold' => "\033[97;1m",
-        'comment' => "\033[33m",
-        'yellow' => "\033[33m",
-        'green' => "\033[32m",
-        'error' => "\033[97;1;41m",
-        'red' => "\033[31m",
-        'info' => "\033[32m",
-        'underlined' => "\033[4m",
-        'underlinedBold' => "\033[4;1m",
-        'underlinedItalic' => "\033[4;3m",
+        'italic' => self::ESC . '[3m',
+        'dark' => self::ESC . '[2m',
+        'bold' => self::ESC . '[1m',
+        'darkItalic' => self::ESC . '[2;3m',
+        'white' => self::ESC . '[97m',
+        'whiteBold' => self::ESC . '[97;1m',
+        'comment' => self::ESC . '[33m',
+        'yellow' => self::ESC . '[33m',
+        'green' => self::ESC . '[32m',
+        'error' => self::ESC . '[97;1;41m',
+        'warning' => self::ESC . '[93m',
+        'red' => self::ESC . '[31m',
+        'info' => self::ESC . '[32m',
+        'underlined' => self::ESC . '[4m',
+        'underlinedBold' => self::ESC . '[4;1m',
+        'underlinedItalic' => self::ESC . '[4;3m',
+
+        'debug' => self::ESC . '[2m',
+        'cyan' => self::ESC . '[36m',
+        'magenta' => self::ESC . '[35m',
+        'black' => self::ESC . '[30m',
+        'blue' => self::ESC . '[34m',
+        'lightGray' => self::ESC . '[37m',
+        'darkGray' => self::ESC . '[90m',
+        'lightRed' => self::ESC . '[91m',
+        'lightGreen' => self::ESC . '[92m',
+        'lightYellow' => self::ESC . '[93m',
+        'lightBlue' => self::ESC . '[94m',
+        'lightMagenta' => self::ESC . '[95m',
+        'lightCyan' => self::ESC . '[96m',
+        'crossed' => self::ESC . '[9m',
     ];
 
-    /** @var Themes */
+    /** @var \AlecRabbit\ConsoleColour\Themes */
     private $colorized;
-    /** @var Themes */
+    /** @var \AlecRabbit\ConsoleColour\Themes */
     private $nonColorized;
 
     /** @test */
@@ -39,6 +58,14 @@ class ThemesTest extends TestCase
         $this->expectException(\BadMethodCallException::class);
         $this->expectExceptionMessage('Unknown method call [AlecRabbit\ConsoleColour\Themes::unknownMethod].');
         $this->assertEquals($text, $this->nonColorized->unknownMethod($text));
+    }
+
+    /** @test */
+    public function none(): void
+    {
+        $text = 'sample';
+        $this->assertEquals($text, $this->nonColorized->none($text));
+        $this->assertEquals($text, $this->colorized->none($text));
     }
 
     /** @test */
@@ -61,7 +88,9 @@ class ThemesTest extends TestCase
         foreach (self::THEMES as $methodName => $theme) {
             $result = $this->colorized->$methodName($text);
             $this->assertEquals(
-                Helper::stripEscape(self::THEMES[$methodName]) . $text . '\033[0m', Helper::stripEscape($result)
+                Helper::stripEscape(self::THEMES[$methodName]) . $text . '\033[0m',
+                Helper::stripEscape($result),
+                $methodName
             );
             $this->assertEquals($text, $this->nonColorized->$methodName($text));
         }